[MT-4437] Group alignment util commands together
