+++ packages/slate-commons/src/commands/index.ts
@@ -1,6 +1,6 @@
+export { getAlignment, toggleAlignment } from './alignment';
 export { findDescendants } from './findDescendants';
 export { focus } from './focus';
-export { getAlignments } from './getAlignments';
 export { getCurrentDomNode } from './getCurrentDomNode';
 export { getCurrentNodeEntry } from './getCurrentNodeEntry';
 export { getCursorPositionInNode } from './getCursorPositionInNode';
@@ -34,5 +34,4 @@ export { removeNode } from './removeNode';
 export { saveSelection } from './saveSelection';
 export { toDomNode } from './toDomNode';
 export { toDomRange } from './toDomRange';
-export { toggleAlignment } from './toggleAlignment';
 export { toggleMark } from './toggleMark';
